+++ atoum/forms/__init__.py
@@ -1,12 +1,13 @@
 from .category import CategoryAdminForm
 from .product import ProductAdminForm
 from .search import GlobalSearchForm
-from .shopping import ShoppingItemInlineForm
+from .shopping import ShoppingAdminForm, ShoppingItemInlineForm
 
 
 __all__ = [
     "CategoryAdminForm",
     "GlobalSearchForm",
     "ProductAdminForm",
+    "ShoppingAdminForm",
     "ShoppingItemInlineForm",
 ]